[KICKS-15][UI][feat] Update data dummy
