+++ frontend/src/components/common/components/ProductCard.css
@@ -9,12 +9,28 @@
   padding: 0;
   overflow: hidden;
   position: relative;
+  transition: transform 0.3s ease, box-shadow 0.3s ease;
+}
+
+.product-card:hover {
+  transform: translateY(-8px);
+}
+
+.product-card__image-container {
+  width: 100%;
+  aspect-ratio: 1 / 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  overflow: hidden;
+  border-radius: 25px;
+  background: #f5f5f5;
 }
 
 .product-card__image {
   width: 100%;
   height: 100%;
-  object-fit: contain;
+  object-fit: cover;
   border-radius: 20px;
   background: transparent;
   display: block;
@@ -36,15 +52,8 @@
   display: flex;
   align-items: center;
   justify-content: center;
-}
-
-.product-card__image {
-  width: 100%;
-  height: 100%;
-  object-fit: contain;
-  border-radius: 20px;
-  background: transparent;
-  display: block;
+  top: 0;
+  left: 0;
 }
 
 .product-card__info {
