+++ frontend/src/components/pages/dashboard/UserManagementPage.jsx
@@ -2,7 +2,7 @@ import React, { useContext, useEffect, useState } from "react";
 import { Button, Pagination } from "antd";
 import TableUsers from "./components/TableUsers";
 import TabHeader from "./components/TabHeader";
-import { getUsers, getTotalUsers } from "./data/dashboardData";
+import { getUsers, getTotalUsers } from "../../../data/mockData";
 import { ActiveTabContext } from "./Dashboard";
 
 const UserManagementPage = () => {
